# website/static/json/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "couic",
        "description": "A lightweight eBPF-powered network filtering solution controllable through a REST API",
        "license": {
            "name": "MIT OR Apache-2.0",
            "identifier": "MIT OR Apache-2.0"
        },
        "version": "0.5.12"
    },
...
    "paths": {
...
        "/v1/stats": {
            "get": {
                "tags": [
                    "stats"
                ],
                "summary": "Get XDP statistics",
                "description": "Returns basic XDP statistics including CIDR counts and packet/byte counters per XDP action.",
                "operationId": "get_stats",
                "responses": {
                    "200": {
                        "description": "Statistics in JSON format",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Stats"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CompositeError": {
                "type": "object",
                "required": [
                    "code",
                    "message",
                    "errors"
                ],
                "properties": {
                    "code": {
                        "$ref": "#/components/schemas/ErrorCode"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "$ref": "#/components/schemas/ErrorDetail"
                        },
                        "propertyNames": {
                            "type": "string"
                        }
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
...
            "ErrorCode": {
                "type": "string",
                "description": "Enum representing error codes used in the API",
                "enum": [
                    "Eprocessing",
                    "Eunauthorized",
                    "Enotfound",
                    "Econflict",
                    "Ebadrequest",
                    "Einvalid",
                    "Einternal",
                    "Enotimplemented"
                ]
            },
            "ErrorDetail": {
                "type": "object",
                "required": [
                    "code",
                    "message"
                ],
                "properties": {
                    "code": {
                        "$ref": "#/components/schemas/ErrorCode"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
...
            "PktStats": {
                "type": "object",
                "description": "Packet and byte counters",
                "required": [
                    "rx_packets",
                    "rx_bytes"
                ],
                "properties": {
                    "rx_packets": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 0,
                        "description": "Number of packets received"
                    },
                    "rx_bytes": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 0,
                        "description": "Number of bytes received"
                    }
                }
            },
...
            "Stats": {
                "type": "object",
                "description": "Basic XDP statistics",
                "required": [
                    "drop_cidr_count",
                    "ignore_cidr_count",
                    "xdp"
                ],
                "properties": {
                    "drop_cidr_count": {
                        "type": "integer",
                        "minimum": 0,
                        "description": "Number of CIDRs in the drop list"
                    },
                    "ignore_cidr_count": {
                        "type": "integer",
                        "minimum": 0,
                        "description": "Number of CIDRs in the ignore list"
                    },
                    "xdp": {
                        "type": "object",
                        "description": "XDP action statistics (packets and bytes per action)",
                        "additionalProperties": {
                            "$ref": "#/components/schemas/PktStats"
                        },
                        "propertyNames": {
                            "type": "string"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Uuidv4",
                "description": "An API key is a token that you provide when making API calls. Include the token in a header parameter called `Authorization`.\nExample: `Authorization: Bearer b03dd133-3827-4793-b0df-37f6ffbdd9a9`"
            }
        }
    },
...
}